+++ Cargo.toml
@@ -1,6 +1,7 @@
 [workspace]
 resolver = "2"
 members = [
+	"bindings",
 	"cli",
 	"sdk",
 	"sdk/src/client/bindings/java/lib/native",
