+++ sdk/tests/mod.rs
@@ -3,8 +3,6 @@
 
 #[cfg(feature = "client")]
 mod client;
-#[cfg(all(feature = "message_interface", feature = "wallet"))]
-mod message_interface;
 #[cfg(feature = "pow")]
 mod pow;
 mod types;
